docs(core): tighten preagg docs and fold exec shim into Connector.ts
