+++ docs/api/core/coordinator.md
@@ -20,7 +20,9 @@ Create a new Mosaic Coordinator to manage all database communication for clients
 * _logger_: The logger to use, defaults to `console`.
 * _cache_: Boolean flag to enable/disable query caching (default `true`).
 * _consolidate_ Boolean flag to enable/disable query consolidation (default `true`).
-* _preagg_: Pre-aggregation options object. The _enabled_ flag (default `true`) determines if pre-aggregation optimizations should be used when possible. The _mode_ option (default `'exec'`) determines how materialized views are created: `'exec'` issues `CREATE TABLE` statements directly, while `'preagg'` sends a `preagg` request and lets the [connector](./connectors) or its server create and name the table. The _schema_ option (default `'mosaic'`) indicates the database schema in which materialized view tables should be created in `'exec'` mode.
+* _preagg_: Pre-aggregation options object. The _enabled_ flag (default `true`) determines if pre-aggregation optimizations should be used when possible. The _mode_ option (default `'exec'`) determines how materialized views are created:
+  * `'exec'` issues `CREATE TABLE` statements directly. The _schema_ option (default `'mosaic'`) indicates the database schema in which the tables are created.
+  * `'preagg'` sends a `preagg` request and lets the [connector](./connectors) or its server create and name the table.
 
 ## databaseConnector
 
@@ -59,6 +61,8 @@ Resets the state of the coordinator. Supports the following _options_:
 
 - _clients_: A Boolean flag (default `true`) indicating if all current clients should be disconnected.
 - _cache_: A Boolean flag (default `true`) indicating if the query cache should be cleared.
+
+When both flags are set, the pre-aggregator is also reset, forgetting any `preagg` table references.
 
 ## exec
