+++ packages/mosaic/core/src/connectors/wasm.ts
@@ -7,9 +7,9 @@ import type {
   PreaggRequest,
   PreaggResponse
 } from './Connector.js';
+import { materializeWithExec } from './Connector.js';
 import * as duckdb from '@duckdb/duckdb-wasm';
 import { decodeIPC } from '../util/decode-ipc.js';
-import { materializeWithExec } from './materialize.js';
 
 interface DuckDBWASMOptions {
   /** Flag to enable logging. */
